+++ source/Private/Invoke-ReportServerSetupAction.ps1
@@ -147,7 +147,7 @@ function Invoke-ReportServerSetupAction
 
         [Parameter(ParameterSetName = 'Install')]
         [Parameter(ParameterSetName = 'Repair')]
-        [ValidateSet('Development', 'Evaluation', 'ExpressAdvanced')]
+        [ValidateSet('Developer', 'Evaluation', 'ExpressAdvanced')]
         [System.String]
         $Edition,
 
@@ -213,6 +213,9 @@ function Invoke-ReportServerSetupAction
 
     Assert-BoundParameter @assertBoundParameters
 
+    # Sensitive values.
+    $sensitiveValue = @()
+
     # Default action is install or upgrade.
     $setupArgument = '/quiet /IAcceptLicenseTerms'
 
@@ -228,6 +231,10 @@ function Invoke-ReportServerSetupAction
     if ($ProductKey)
     {
         $setupArgument += ' /PID={0}' -f $ProductKey
+
+        $sensitiveValue += @(
+            $ProductKey
+        )
     }
 
     if ($EditionUpgrade.IsPresent)
@@ -237,7 +244,13 @@ function Invoke-ReportServerSetupAction
 
     if ($Edition)
     {
-        $setupArgument += ' /Edition={0}' -f $Edition
+        $editionMap = @{
+            Developer = 'Dev'
+            Evaluation = 'Eval'
+            ExpressAdvanced = 'ExprAdv'
+        }
+
+        $setupArgument += ' /Edition={0}' -f $editionMap.$Edition
     }
 
     if ($LogPath)
@@ -256,11 +269,6 @@ function Invoke-ReportServerSetupAction
     }
 
     $verboseSetupArgument = $setupArgument
-
-    # Sensitive values.
-    $sensitiveValue = @(
-        $ProductKey
-    )
 
     # Obfuscate sensitive values.
     foreach ($currentSensitiveValue in $sensitiveValue)